add script to generate docfy markdown
