--- ui/metadata.json
@@ -1,6 +1,0 @@
-{
-  "versions": {
-    "main": "https://vault-storybook.vercel.app/"
-  }
-}
-
